exam/requirements.txt
backcall==0.1.0
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-filter==2.1.0
django-rest-auth==0.9.5
django-url-filter==0.3.12
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
enum-compat==0.0.2
idna==2.8
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.1
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7

